requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
darkdetect==0.7.1
docopt==0.6.2
idna==3.7
MouseInfo==0.1.3
numpy==1.24.4
opencv-python==4.9.0.80
pillow==10.3.0
pipreqs==0.4.13
psutil==5.9.5
PyAudio==0.2.14
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.6
pyperclip==1.8.2
pyqtdarktheme==2.1.0
PyRect==0.2.0
PyScreeze==0.1.30
pytweening==1.2.0
pywin32==306
requests==2.31.0
six==1.16.0
urllib3==2.2.1
yarg==0.1.9

PyQt5~=5.15.10